+++ YoursMine/YoursMine/니꺼사기/BuyPageViewController.swift
@@ -16,7 +16,6 @@ class BuyPageViewController: UIPageViewController, UIPageViewControllerDataSourc
 
     var pageDelegate: PageIndexDelegate?
     var pendingPage: Int?
-    let identifiers: NSArray = ["AllVC", "DailyVC", "SupportVC"]
 
     lazy var VCArray: [UIViewController] = {
         return [self.VCInstance(name: ProductListController.name),
@@ -27,7 +26,6 @@ class BuyPageViewController: UIPageViewController, UIPageViewControllerDataSourc
 
     required init?(coder aDecoder: NSCoder) {
         super.init(transitionStyle: .scroll, navigationOrientation: .horizontal, options: nil)
-
     }
 
     private func VCInstance(name: String) -> UIViewController {
@@ -46,12 +44,29 @@ class BuyPageViewController: UIPageViewController, UIPageViewControllerDataSourc
     }
     
     func pageViewController(_ pageViewController: UIPageViewController, viewControllerBefore viewController: UIViewController) -> UIViewController? {
-        <#code#>
+        guard let vcIndex = VCArray.firstIndex(of: viewController) else { return nil }
+        let previousIndex = vcIndex - 1
+        if previousIndex < 0 {
+            return VCArray.last
+        } else {
+            return VCArray[previousIndex]
+        }
+        
+        self.pageDelegate?.selectMenuItem(pageIndex: vcIndex)
     }
     
     func pageViewController(_ pageViewController: UIPageViewController, viewControllerAfter viewController: UIViewController) -> UIViewController? {
-        <#code#>
+        guard let vcIndex = VCArray.firstIndex(of: viewController) else {
+            return nil
+        }
+        let nextIndex = vcIndex + 1
+        if nextIndex >= VCArray.count {
+            return VCArray.first
+        } else {
+            return VCArray[nextIndex]
+        }
+        
+        self.pageDelegate?.selectMenuItem(pageIndex: vcIndex)
     }
-    
 
 }
