--- YoursMine/YoursMine/니꺼사기/Cells/ProductSellingCell.xib
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="16097.2" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_1" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16087"/>
-        <capability name="Named colors" minToolsVersion="9.0"/>
-        <capability name="collection view cell content view" minToolsVersion="11.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="ProductSellingCell" id="Z0d-td-oVA" customClass="ProductSellingCell" customModule="YoursMine" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="601" height="727"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <collectionViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" insetsLayoutMarginsFromSafeArea="NO" id="tXa-wi-xVg">
-                <rect key="frame" x="0.0" y="0.0" width="601" height="727"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="SHI-bq-syd">
-                        <rect key="frame" x="0.0" y="0.0" width="601" height="727"/>
-                        <subviews>
-                            <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" showsHorizontalScrollIndicator="NO" style="plain" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="6X4-RE-mvR">
-                                <rect key="frame" x="0.0" y="0.0" width="601" height="727"/>
-                                <color key="backgroundColor" name="LightYellowColor"/>
-                            </tableView>
-                        </subviews>
-                        <color key="backgroundColor" name="LightYellowColor"/>
-                        <constraints>
-                            <constraint firstAttribute="bottom" secondItem="6X4-RE-mvR" secondAttribute="bottom" id="NqG-r7-hDD"/>
-                            <constraint firstItem="6X4-RE-mvR" firstAttribute="top" secondItem="SHI-bq-syd" secondAttribute="top" id="o8O-TB-ld7"/>
-                            <constraint firstItem="6X4-RE-mvR" firstAttribute="leading" secondItem="SHI-bq-syd" secondAttribute="leading" id="th7-pU-6g5"/>
-                            <constraint firstAttribute="trailing" secondItem="6X4-RE-mvR" secondAttribute="trailing" id="x3r-sw-c9s"/>
-                        </constraints>
-                    </view>
-                </subviews>
-                <constraints>
-                    <constraint firstAttribute="trailing" secondItem="SHI-bq-syd" secondAttribute="trailing" id="1uj-R5-MAg"/>
-                    <constraint firstItem="SHI-bq-syd" firstAttribute="top" secondItem="tXa-wi-xVg" secondAttribute="top" id="Zdh-p4-VeF"/>
-                    <constraint firstItem="SHI-bq-syd" firstAttribute="leading" secondItem="tXa-wi-xVg" secondAttribute="leading" id="kqC-IJ-k3G"/>
-                    <constraint firstAttribute="bottom" secondItem="SHI-bq-syd" secondAttribute="bottom" id="vid-rU-k70"/>
-                </constraints>
-            </collectionViewCellContentView>
-            <size key="customSize" width="601" height="727"/>
-            <connections>
-                <outlet property="tableView" destination="6X4-RE-mvR" id="aQQ-Xc-TPm"/>
-            </connections>
-            <point key="canvasLocation" x="105.07246376811595" y="202.56696428571428"/>
-        </collectionViewCell>
-    </objects>
-    <resources>
-        <namedColor name="LightYellowColor">
-            <color red="1" green="0.92941176470588238" blue="0.7686274509803922" alpha="0.40000000596046448" colorSpace="custom" customColorSpace="sRGB"/>
-        </namedColor>
-    </resources>
-</document>
