+++ YoursMine/YoursMine/니꺼사기/물품구매/AuctionListController.swift
@@ -11,10 +11,28 @@ import UIKit
 class AuctionListController: UIViewController {
     static let name = "AuctionListController"
     
+    @IBOutlet weak var tableView: UITableView!
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        // Do any additional setup after loading the view.
+        tableView.delegate = self
+        tableView.dataSource = self
+        let itemListCell = UINib(nibName: ItemListCell.name, bundle: nil)
+        tableView.register(itemListCell, forCellReuseIdentifier: ItemListCell.name)
+        tableView.estimatedRowHeight = 110
     }
     
+}
+extension AuctionListController: UITableViewDataSource, UITableViewDelegate {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return 5
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        if let itemListCell = tableView.dequeueReusableCell(withIdentifier: ItemListCell.name, for: indexPath) as? ItemListCell {
+            return itemListCell
+        }
+        
+        return UITableViewCell()
+    }
 }
