+++ YoursMine/YoursMine/니꺼사기/물품구매/GroupBuyingListController.swift
@@ -10,22 +10,29 @@ import UIKit
 
 class GroupBuyingListController: UIViewController {
 
+    @IBOutlet weak var tableView: UITableView!
     static let name = "GroupBuyingListController"
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        // Do any additional setup after loading the view.
+        tableView.delegate = self
+        tableView.dataSource = self
+        let itemListCell = UINib(nibName: ItemListCell.name, bundle: nil)
+        tableView.register(itemListCell, forCellReuseIdentifier: ItemListCell.name)
+        tableView.estimatedRowHeight = 110
+    }
+}
+
+extension GroupBuyingListController: UITableViewDataSource, UITableViewDelegate {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return 5
     }
     
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        if let itemListCell = tableView.dequeueReusableCell(withIdentifier: ItemListCell.name, for: indexPath) as? ItemListCell {
+            return itemListCell
+        }
+        
+        return UITableViewCell()
     }
-    */
-
 }
